% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_method.R
\name{anova.wizirt}
\alias{anova.wizirt}
\title{ANOVA method for wizirt objects}
\usage{
anova.wizirt(x, y = NULL)
}
\arguments{
\item{x}{An object of class wizirt (i.e. from fit_wizirt).}

\item{y}{An object of class wizirt (i.e. from fit_wizirt).}
}
\description{
Function to make ANOVA comparisons for wizirt models. If only one model
}
