% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip_style_functions.R
\name{fit_wizirt}
\alias{fit_wizirt}
\title{Run the irt model you described}
\usage{
fit_wizirt(object, data, control = parsnip:::control_parsnip(), ...)
}
\arguments{
\item{object}{An object from set_engine()}

\item{data}{An Person x Items matrix or dataframe of dichotomous response values (e.g. correct/incorrect). Rows are persons and columns are items, one row per person, one column per item. No other information allowed.}
}
\description{
fit_wizirt() runs a model that has been built with irt() and set_engine().
}
