% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip_style_functions.R
\name{irt}
\alias{irt}
\title{Describe the model you will run}
\usage{
irt(
  mode = "regression",
  item_type = NULL,
  irt_pars = TRUE,
  rownames = NULL,
  tol = 1e-05
)
}
\arguments{
\item{mode}{Must be 'regression' currently.}

\item{item_type}{Character. Must be one of 'Rasch', '1PL', '2PL' or '3PL'.}

\item{irt_pars}{Logical. Should the traditional IRT parametrization be used? If false the slope-intercept form is used.}

\item{rownames}{Optional unique row IDs for the data (i.e. examinee IDs). If omitted, uses 1:nrow(data).}

\item{tol}{Numeric. Convergence criterion. Currently only implemented when engine is mirt.}
}
\description{
irt() is used to describe a model, it must be used in conjunction with set_engine(), and fit_wizirt().
}
