% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_assume.R
\name{irt_assume}
\alias{irt_assume}
\title{A function for checking the assumptions of IRT.}
\usage{
irt_assume(wizirt_fit, verbose = FALSE)
}
\arguments{
\item{wizirt_fit}{An object from fit_wizirt}

\item{verbose}{Logical. Should the names of the other packages used in this function be printed? Default is FALSE.
#' @return A list with tibbles of data in them.}
}
\description{
A function for checking the assumptions of IRT.
}
