% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_item_fit.R
\name{irt_item_fit}
\alias{irt_item_fit}
\title{A function used to get additional item-level information, such as ICC probabilities, and item information.}
\usage{
irt_item_fit(wizirt_fit, stats = c("X2"))
}
\arguments{
\item{wizirt_fit}{An object coming from the fit_wizirt function.}

\item{stats}{A character or character string identifying item-level fit measures. Must be at least one of c('Zh', 'X2', 'G2', 'infit'). More are coming very soon.  Default is 'X2'.}
}
\value{
A list with item_stats (item-level statistics), item_information (data for item and test information), and item_probabilities (data for ICC curves).
}
\description{
A function used to get additional item-level information, such as ICC probabilities, and item information.
}
\examples{
ifa <- wizirt2:::irt_item_fit(my_model)
}
