% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_model_pfa.R
\name{irt_model_pfa}
\alias{irt_model_pfa}
\title{A function to explore reasons for person misfit.}
\usage{
irt_model_pfa(wizirt_fit, pfa = NULL, predictors = NULL, bins = 5)
}
\arguments{
\item{wizirt_fit}{An object of class wizirt (i.e. from fit_wizirt)}

\item{pfa}{An object of class wizirt_pfa (i.e. from irt_person_fit)}

\item{predictors}{Either a matrix or dataframe with a row per person and a column per predictor.}

\item{bins}{Numeric. How many bins to break items into. Sorted by difficulty then broken into bins. If items cannot fit equally into bins, the bins of easier items will take extras.}
}
\description{
A function to explore reasons for person misfit.
}
