% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_report.R
\name{irt_report}
\alias{irt_report}
\title{Create a general IRT report for a test}
\usage{
irt_report(
  data,
  report_type = "html",
  title = "My Report",
  author = "",
  note = "",
  engine = "mirt",
  item_type = "Rasch",
  rownames = NULL,
  tol = 1e-05,
  irt_pars = TRUE,
  ifa_stats = c("X2"),
  pfa_predictors = NULL,
  pfa_bins = 5,
  pfa_stats = c("Ht")
)
}
\arguments{
\item{item_type}{Character. Must be one of 'Rasch', '1PL', '2PL' or '3PL'.}

\item{rownames}{Optional unique row IDs for the data (i.e. examinee IDs). If omitted, uses 1:nrow(data).}

\item{tol}{Numeric. Convergence criterion. Currently only implemented when engine is mirt.}

\item{irt_pars}{Logical. Should the traditional IRT parametrization be used? If false the slope-intercept form is used.}

\item{ifa_stats}{A character or character string identifying item-level fit measures. Must be at least one of c('Zh', 'X2', 'G2', 'infit'). More are coming very soon.  Default is 'X2'.}

\item{pfa_predictors}{Either a matrix or dataframe with a row per person and a column per predictor.}

\item{pfa_bins}{Numeric. How many bins to break items into for pfa multilevel modeling. Sorted by difficulty then broken into bins. If items cannot fit equally into bins, the bins of easier items will take extras.}

\item{pfa_stats}{A character or character string identifying person-level fit measures. Default is "Ht". All of the stats in PerFit should be available. Let me know if any don't work. for more information.}

\item{mode}{Must be 'regression' currently.}
}
\description{
irt_report
}
