% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot_wrap}
\alias{plot_wrap}
\title{A funirction for handling large numbers of students and items in plots.}
\usage{
plot_wrap(object, type, items_per = NULL, persons_per = NULL)
}
\arguments{
\item{object}{An object of class wizirt_fit.}

\item{type}{Character. Currently, can be 'obs', 'trace', 'info', 'resid', 'stand', 'tinfo', 'theta', 'diff', 'theta_diff', 'np_prf', 'ld', or 'ld_pairs'.}

\item{items_per}{Numeric. How many items to plot per page?}

\item{persons_per}{Numeric. How many persons to plot per page?}
}
\description{
This function helps to break the plots into several plots when many items or persons are present.
}
