% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.wizirt_assume}
\alias{print.wizirt_assume}
\title{Print method for wizirt assumption objects}
\usage{
\method{print}{wizirt_assume}(x, type = "all")
}
\arguments{
\item{x}{An object exported from irt_assume()}

\item{type}{Character. One of 'all', 'ld', 'unid', 'rel', 'abs'. Default is 'all'.}
}
\description{
Print method for wizirt assumption objects
}
