% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.wizirt_ifa}
\alias{print.wizirt_ifa}
\title{Print method for wizirt item-fit objects}
\usage{
\method{print}{wizirt_ifa}(x)
}
\arguments{
\item{x}{An object exported from irt_item_fit()}
}
\description{
Print method for wizirt item-fit objects
}
