% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{set_engine}
\alias{set_mode}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{parsnip}{\code{\link[parsnip]{set_engine}}, \code{\link[parsnip:set_args]{set_mode}}}
}}

