% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip_style_functions.R
\name{wizirt}
\alias{wizirt}
\title{Estimate an IRT model}
\usage{
wizirt(
  data,
  rownames = NULL,
  item_type = "Rasch",
  engine = "mirt",
  irt_pars = TRUE,
  tol = 1e-05
)
}
\arguments{
\item{data}{An Person x Items matrix or dataframe of dichotomous response values (e.g. correct/incorrect). Rows are persons and columns are items, one row per person, one column per item. No other information allowed.}

\item{rownames}{Optional unique row IDs for the data (i.e. examinee IDs). If omitted, uses 1:nrow(data).}

\item{item_type}{Character. Must be one of 'Rasch', '1PL', '2PL' or '3PL'.}

\item{engine}{Character. Currently supported engines are 'mirt' and 'ltm' for Rasch, 1PL, 2PL, and 3PL models. 'eRm' is supported for Rasch models only.}

\item{irt_pars}{Logical. Should the traditional IRT parametrization be used? If false the slope-intercept form is used.}

\item{tol}{Numeric. Convergence criterion. Currently only implemented when engine is mirt.}
}
\description{
Estimate an IRT model using various engines
}
