% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsnip_style_functions.R
\name{fit_wizirt}
\alias{fit_wizirt}
\title{Run the irt model you described}
\usage{
fit_wizirt(object, data, control = parsnip:::control_parsnip(), ...)
}
\arguments{
\item{object}{An object from set_engine()}

\item{data}{An Person x Items matrix or dataframe of dichotomous response values (e.g. correct/incorrect). Rows are persons and columns are items, one row per person, one column per item. No other information allowed.}
}
\value{
Returns a list of class wizirt.
\emph{spec} is a list of information for the parsnip backend. printing spec prints a summary of the model run.
\emph{elapsed} contains the time it took the model to run.
\emph{fit} contains the model information:
\itemize{
\item \emph{data} is the data passed to the model
\item \emph{model} contains model fit information, including:
}
\itemize{
\item \emph{engine} a list with values pkg (the package used for estimation), ver (pkg version), func (functiong used from pkg), and call (call made to pkg)
\item \emph{n_factors} the number of factors estimated
\item \emph{item_type} the item type passed to wizirt (Rasch, 1PL, 2PL, or 3PL).
}
\itemize{
\item \emph{estimation} a list with information related to convergence. \emph{convergence} a T/F value of whether the model converged, \emph{method} the estimation method, \emph{criteria} the convergence criteria, \emph{iterations} the number of iterations it took for the model to converge, \emph{log_lik} the loglikelihood at convergence, \emph{abs_fit} the absolute fit of the model, \emph{df} the number of parameters estimated.
\item \emph{parameters} a list of estimated parameters. \emph{coefficients} is a data frame of estimated item-statistics and \emph{persons} is a data frame of estimated person statistics.
\item \emph{original_object} is the object returned from the engine.
}
}
\description{
fit_wizirt() runs a model that has been built with irt() and set_engine().
}
\examples{
data("responses")
my_model <- irt(item_type = "Rasch") \%>\%
  set_engine(engine = "mirt") \%>\%
  fit_wizirt(data = responses[, -1])
}
