% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irt_assume.R
\name{irt_assume}
\alias{irt_assume}
\title{A function for checking the assumptions of IRT.}
\usage{
irt_assume(wizirt_fit, verbose = FALSE)
}
\arguments{
\item{wizirt_fit}{An object from fit_wizirt}

\item{verbose}{Logical. Should the names of the other packages used in this function be printed? Default is FALSE.}
}
\value{
A list with tibbles of data in them.
}
\description{
This function checks the assumptions of IRT. For unidimensionality the DETECT, RATIO, and ASSI statistics are calculated using \code{\link[sirt:conf.detect]{sirt::conf.detect()}}.
For local dependence, the standardized LD statistic is calculated using \code{\link[mirt:residuals-method]{mirt::residuals-method()}} and the correlations are created using \code{\link[ltm:rcor.test]{ltm::rcor.test()}}.
Relative fit calculates BIC, AIC, and several others using code from \code{\link[mirt:anova-method]{mirt::anova-method()}}.
Absolute fit is calculated only for models run with the mirt engine, using \code{\link[mirt:M2]{mirt::M2()}}.
See the documentation on those functions for more information on calculation.
}
\examples{
data("responses")
my_model <- wizirt(data = responses[,-1])
assumptions <- irt_assume(my_model)
print(assumptions, type = 'rel')
print(assumptions, type = 'abs')
print(assumptions, type = 'unid')
print(assumptions, type = 'ld')
}
