% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_methods.R
\name{plot.wizirt}
\alias{plot.wizirt}
\title{Plot method for wizirt objects}
\usage{
\method{plot}{wizirt}(
  wizirt_fit,
  type = "tinfo",
  items = NULL,
  persons = NULL,
  facets = TRUE,
  quads = 10,
  return_data = FALSE,
  pfa = NULL,
  ifa = NULL
)
}
\arguments{
\item{type}{Character. Currently, can be 'obs', 'trace', 'info', 'resid', 'stand', 'tinfo', 'theta', 'diff', 'theta_diff', 'np_prf', 'ld', or 'ld_pairs'.}

\item{items}{Which items to plot? Either a numeric vector of item positions in the column names of the data, or a vector of the item names to plot. If nothing is specifed all items will be plotted.}

\item{persons}{Which persons to plot? Either a numeric vector of person positions in the row names of the data, or a vector of the person ids to plot. If nothing is specifed all persons will be plotted.}

\item{facets}{Logical. Should the plots be faceted? Default is TRUE.}

\item{quads}{Numeric. For plots using residuals (i.e. resid, stand). How many quantiles should the data be broken into?}

\item{return_data}{Logical. Should the plot data be returned. If TRUE returns a list with the plot and the data.}

\item{pfa}{An object from irt_person_fit(). If omitted, irt_person_fit() is called within the function with the default settings (only for np_prf)}

\item{ifa}{An object from irt_item_fit(). If omitted, irt_item_fit() is called within the function with the default settings (only for tinfo)}
}
\description{
All plots in wizirt use ggplot2 as a backend. Type 'theta_diff' also uses gghalves. For the item-focused plots, the types can be included in the same string to overlay the plots (e.g. resid_trace).
For large numbers of persons and/or items, you may want to consider limiting the number printed at a time on some plots.
}
\examples{
data("responses")
my_model <- wizirt(data = responses[,-1])
plot(my_model, type = 'tinfo')
plot(my_model, type = 'theta')
plot(my_model, type = 'diff')
plot(my_model, type = 'theta_diff')
plot(my_model, type = 'obs', items = 1:4)
plot(my_model, type = 'trace', items = 1:4)
plot(my_model, type = 'trace', items = 1:4, facets = FALSE)
plot(my_model, type = 'resid', items = 1:4)
plot(my_model, type = 'stand', items = 1:4)
plot(my_model, type = 'info', items = 1:4)
plot(my_model, type = 'resid_trace', items = 1:4)
plot(my_model, type = 'np_prf', facets = T, persons = 1:4)
plot(my_model, type = 'np_prf', facets = F)
}
