% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.wizirt_pfa}
\alias{print.wizirt_pfa}
\title{Print method for wizirt item-fit objects}
\usage{
\method{print}{wizirt_pfa}(x, patterns = FALSE)
}
\arguments{
\item{x}{An object exported from irt_person_fit()}

\item{patterns}{Logical. Should the response patterns be printed as well?}
}
\description{
Print method for wizirt item-fit objects
}
