% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{set_engine}
\alias{set_mode}
\title{Objects exported from other packages}
\arguments{
\item{engine}{for set_engine(). Character. Currently supported engines are 'mirt' and 'ltm' for Rasch, 1PL, 2PL, and 3PL models. 'eRm' is supported for Rasch models only.}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{parsnip}{\code{\link[parsnip]{set_engine}}, \code{\link[parsnip:set_args]{set_mode}}}
}}

